requirements.txt
streamlit
tensorflow
matplotlib
pandas
python-dotenv
kaggle
pillow
numpy
requests
streamlit-lottie
pyttsx3